Add integrated tests for search result ordering
Test support for:

 - Sorting by "updated" and "created" fields
 - Default sort ordering
 - What happens when an invalid sort field is specified

As far as I can tell we only ever sort annotations in chronological
order of last update in h and the client (ie. sort by updated,
descending). However the search API currently supports sorting by any
field which is indexed in Elasticsearch. We might be able to drop that
flexibility if it is unused.